Worked on extracting config data in other files
You can now get the values of the config file anywhere in the project as long as you have access to config.rs.
Error: cannot return value referencing temporary value
This error occurs when trying to use the config().writing.version in the clap version section.
No idea on how to fix this yet