Started working on settings
No fix yet on the clap problem so I commented it out and started focusing on the settings for now.

I created a settings.rs file currently with the public function version which is finished and with update which is currently under development.

No errors are known to me at this time.